+++ e2e/uploads/sanPathSelect.spec.ts
@@ -15,7 +15,7 @@ test('can enter a SAN path and output is rendered correctly', async ({ page }) =
   })
 
   // assert details about path are in rendered distribution option
-  const downloadOutput = await page.textContent('#download-1-output pre')
+  const downloadOutput = await page.textContent('#downloads-output pre')
   expect(downloadOutput).toContain(`"href": ""`)
   expect(downloadOutput).toContain(
     '"href": "https://www.iana.org/assignments/media-types/image/png"'