+++ src/components/sections/elements/Download.vue
@@ -2,17 +2,17 @@
 import { computed, type ComputedRef, type PropType, ref, watch } from 'vue'
 
 import { ResourceType } from '@/types/enum'
-import type { Licence, Upload as UploadT } from '@/types/app'
+import type { DistributionOptionIndexed, Licence, Upload as UploadT } from '@/types/app'
 import type { DistributionOption, PointOfContact as IsoContact } from '@/types/iso'
 import { createDistributor, createDownloadDistributionOption } from '@/lib/distribution'
 
-import Output from '@/components/bases/Output.vue'
 import SubSectionBorder from '@/components/bases/SubSectionBorder.vue'
 import Upload from '@/components/bases/Upload.vue'
+import Button from '@/components/bases/Button.vue'
 
 const props = defineProps({
   index: {
-    type: Number,
+    type: String,
     required: true,
   },
   fileIdentifier: {
@@ -30,8 +30,13 @@ const props = defineProps({
 })
 
 const emit = defineEmits<{
-  'update:isoDistributionOption': [id: DistributionOption]
+  destroy: [id: string]
+  'update:distributionOptionIndexed': [id: DistributionOptionIndexed]
 }>()
+
+const destroy = () => {
+  emit('destroy', props.index)
+}
 
 let upload = ref<UploadT | undefined>(undefined)
 
@@ -50,28 +55,31 @@ let distributionOption: ComputedRef<DistributionOption | undefined> = computed((
   )
 })
 
+let distributionOptionIndexed: ComputedRef<DistributionOptionIndexed> = computed(() => {
+  return {
+    index: props.index,
+    distributionOption: distributionOption.value,
+  }
+})
+
 watch(
   () => upload.value,
   () => {
     if (distributionOption.value) {
-      emit('update:isoDistributionOption', distributionOption.value)
+      emit('update:distributionOptionIndexed', distributionOptionIndexed.value)
     }
   }
 )
 </script>
 
 <template>
-  <SubSectionBorder :id="'download-' + index" class="space-y-2">
+  <SubSectionBorder :id="'download-' + index" class="flex space-x-4">
     <Upload
       :context="'download'"
       :identifier="index"
       :file-identifier="fileIdentifier"
       @update:upload="(event: UploadT) => (upload = event)"
     ></Upload>
-    <Output
-      v-if="distributionOption"
-      :id="'download-' + index + '-output'"
-      :data="distributionOption"
-    ></Output>
+    <Button :id="'download-' + index + '-destroy'" @click="destroy()">✖️</Button>
   </SubSectionBorder>
 </template>